+++ data/normalized/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Latest Microsoft Security News Changes
 
-Generated: 2026-05-03T07:58:07.171Z
+Generated: 2026-05-04T08:26:01.487Z
 
 - 2026-07-01 • Microsoft Sentinel • Action required • [Updated] Call to action: update automation by July 1, 2026 - Account Name is now consistently the UPN prefix for analytics rule alerts
 - 2026-07-01 • Microsoft Sentinel • Action required • Call to action: update queries and automation by July 1, 2026 - standardized account entity naming in incidents and alerts
@@ -9,6 +9,7 @@ Generated: 2026-05-03T07:58:07.171Z
 - 2026-06-15 • Microsoft Sentinel • Action required • Call to action: update older Microsoft Sentinel content as code (Sentinel repositories) API versions before June 15, 2026
 - 2026-06-01 • Microsoft Entra • Action required • General Availability - Sign-in with username/alias
 - 2026-06-01 • Microsoft Sentinel • Action required • Required action
+- 2026-05-03 • Microsoft Defender for Cloud • GA • General availability of Microsoft Defender for Cloud and GitHub Advanced Security integration
 - 2026-05-01 • Microsoft Entra • Deprecation • Plan for change – Agent Registry consolidation into Microsoft Agent 365
 - 2026-05-01 • Microsoft Purview • GA • General availability (GA)
 - 2026-04-30 • Microsoft Defender for Cloud • GA • Container runtime anti-malware detection and blocking is now generally available in Defender for Containers for Azure Kubernetes Service (AKS), Amazon Elastic Kubernetes Service (EKS), and Google Kubernetes Engine (GKE).
@@ -21,4 +22,3 @@ Generated: 2026-05-03T07:58:07.171Z
 - 2026-04-01 • Microsoft Defender for Endpoint • Preview • You can now view the current status of automatic attack disruption and predictive shielding actions related to a specific incident. You view this data in the
 - 2026-04-01 • Microsoft Defender for Endpoint • Ikke oppgitt • Release version 30.126032.0000.0 released: See
 - 2026-04-01 • Microsoft Defender for Endpoint • Ikke oppgitt • Release version 20.126032.16.0 released: See
-- 2026-04-01 • Microsoft Defender for Endpoint • Ikke oppgitt • Release version 4.18.25040.1 released: See